Rename inlineAndCompile -> recursiveModuleCompile
Rename compileToSingleFile -> recursiveModuleCompileToSingleFile.
Rename compileToDirectory -> forEachRecursiveModuleCompile

Remove compiler.js and move its content into recursiveModuleCompile
