+++ src/node/api.js
@@ -21,6 +21,8 @@
 
 var path = require('path');
 var traceur = require('./traceur.js');
+var util = require('./file-util.js');
+var writeFile = util.writeFile;
 
 var Compiler = traceur.Compiler;
 
@@ -70,6 +72,26 @@ function moduleToAmd(content) {
       Compiler.amdOptions({}));
 }
 
+function treeToString(input) {
+  return new NodeCompiler().treeToString(input);
+}
+
+function getSourceMapFileName(name) {
+  return name.replace(/\.js$/, '.map');
+}
+
+function writeCompiledCodeToFile(compiledCode, filename, sourcemap) {
+  var sourceMapFilePath
+  if (sourcemap) {
+    sourceMapFilePath = getSourceMapFileName(filename);
+    compiledCode += '\n//# sourceMappingURL=' +
+        path.basename(sourceMapFilePath) + '\n';
+  }
+  writeFile(filename, compiledCode);
+  if (sourcemap)
+    writeFile(sourceMapFilePath, sourcemap);
+}
+
 // The absolute path to traceur-runtime.js -- the file that should be executed
 // if you want to run Traceur-compiled scripts when the compiler isn't present.
 var RUNTIME_PATH = path.join(__dirname, '../../bin/traceur-runtime.js');
@@ -80,5 +102,7 @@ module.exports = {
   compile: compile,
   moduleToCommonJS: moduleToCommonJS,
   moduleToAmd: moduleToAmd,
+  treeToString: treeToString,
+  writeCompiledCodeToFile: writeCompiledCodeToFile,
   RUNTIME_PATH: RUNTIME_PATH
 };
